feat(activity): full osgi implementation for activity module

+++ fitpal-osgi-app/fitpal-api/src/main/java/com/fitpal/api/User.java
@@ -21,9 +21,9 @@ public class User {
     private boolean deactivated;
 
     // Daily Targets
-    private Double dailyTargetCalorie;
+    private Integer dailyTargetCalorie;
     private Integer dailyTargetSteps;
-    private Double dailyTargetActivity;
+    private Integer dailyTargetActivity;
 
     // Complex Type
     private List<FavouriteFood> favouriteFood = new ArrayList<>();
@@ -76,14 +76,14 @@ public class User {
     public boolean isDeactivated() { return deactivated; }
     public void setDeactivated(boolean deactivated) { this.deactivated = deactivated; }
 
-    public Double getDailyTargetCalorie() { return dailyTargetCalorie; }
-    public void setDailyTargetCalorie(Double dailyTargetCalorie) { this.dailyTargetCalorie = dailyTargetCalorie; }
+    public Integer getDailyTargetCalorie() { return dailyTargetCalorie; }
+    public void setDailyTargetCalorie(Integer dailyTargetCalorie) { this.dailyTargetCalorie = dailyTargetCalorie; }
 
     public Integer getDailyTargetSteps() { return dailyTargetSteps; }
     public void setDailyTargetSteps(Integer dailyTargetSteps) { this.dailyTargetSteps = dailyTargetSteps; }
 
-    public Double getDailyTargetActivity() { return dailyTargetActivity; }
-    public void setDailyTargetActivity(Double dailyTargetActivity) { this.dailyTargetActivity = dailyTargetActivity; }
+    public Integer getDailyTargetActivity() { return dailyTargetActivity; }
+    public void setDailyTargetActivity(Integer dailyTargetActivity) { this.dailyTargetActivity = dailyTargetActivity; }
 
     public List<FavouriteFood> getFavouriteFoods() { return favouriteFood; }
     public void setFavouriteFoods(List<FavouriteFood> favouriteFoods) { this.favouriteFood = favouriteFoods; }
